system and APIC .h/.cpp: type mistakes
main.cpp: removing deadspace

LAPIC.cpp/.h: LAPIC can be written to, debugging tools are not helpful in this case, since it is not real memory, the memory mapping of the register always shows 0.
These registers must be written to using a full write of 32 bits and, specifically the SPV, must not affect lowest 4 bits.

IDT.s/.h/.cpp Specifying the number of IDT stubs in one place.

type mistake

+++ include/APIC.cpp
@@ -15,10 +15,10 @@
 
 uintptr_t get_local_apic_base_addr()
 {
-    uintptr_t eax;
-    uintptr_t edx;
+    u32 eax;
+    u32 edx;
     cpu_get_MSR(IA32_APIC_BASE_MSR, &eax, &edx);
-    return (eax & 0xfffff000);
+    return (static_cast<uintptr_t>(eax) & 0xfffff000);
 }
 
 void set_local_apic_base_addr(const uintptr_t local_apic_base_addr)
